[Silver III] Title: 주유소, Time: 32 ms, Memory: 2556 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.cc
@@ -2,23 +2,23 @@
 #include <stdio.h>
 
 int main() {
-	int N;
-	scanf("%d", &N);
-	int d[100000];
-	int c[100000];
-	for (int i = 0; i < N - 1; i++) {
-		scanf("%d", &d[i]);
+	long long N;
+	scanf("%lld", &N);
+	long long d[100000];
+	long long c[100000];
+	for (long long i = 0; i < N - 1; i++) {
+		scanf("%lld", &d[i]);
 	}
-	for (int i = 0; i < N; i++) {
-		scanf("%d", &c[i]);
+	for (long long i = 0; i < N; i++) {
+		scanf("%lld", &c[i]);
 	}
-	int m = c[0] * d[0];
-	int c1 = c[0];
-	for (int i = 1; i < N - 1; i++) {
+	long long m = c[0] * d[0];
+	long long c1 = c[0];
+	for (long long i = 1; i < N - 1; i++) {
 		if (c1 > c[i]) {
 			c1 = c[i];
 		}
 		m += c1 * d[i];
 	}
-	printf("%d", m);
+	printf("%lld", m);
 }
